Partiel CloudSecurityIot/Device-Monitoring-v2/requirements.txt
fastapi==0.115.6
pymongo==4.10.1
requests==2.32.3
python-dotenv==1.0.1
paho-mqtt==1.6.1
python-jose==3.3.0
uvicorn[standard]==0.32.1
python-socketio==5.11.0
websocket-client==1.8.0
prometheus-fastapi-instrumentator==7.0.0
redis>=4.2.0
